requirements.txt
fastavro==0.21.23
google-cloud-pubsub
confluent-kafka[avro] 
apache-beam[gcp]>=2.16.0
